feat: add IniFunctionRegistry
It manages the mapping from function names to IniFunctions.
It is stored in the IniContext.
Now builtins and functions are not stored in the frame but in the function registry.

Removed uncessary instanciation arguments in Builtins functions.
Updated lookup in Invocation and registration in Function.
Now, Print and Println use the context to get the PrintStream.