feat: make some nodes Wrappable
Added @GenerateWrapper and createWrapper to the nodes
They use an empty constructor

In the future, maybe find a way to remove the need to have them wrappable.
This can be done by making the fields more generic
The need for them to be wrappable comes from the fact that the nodes are specificaly
referenced in the fields of other nodes.